requirements.txt
fastapi[all]~=0.99
sqlalchemy~=2.0
python-dotenv~=1.0
pyjwt~=2.8
passlib[bcrypt]~=1.7
apscheduler~=3.10
databases~=0.5
psycopg2-binary
uvicorn[standard]~=0.23
